requirements.txt
Django==4.0.1
django-environ==0.8.1
psycopg2-binary=2.9.3
djangorestframework==3.13.1
